--- src/main/kotlin/racing/Application.kt
@@ -1,7 +1,0 @@
-package racing
-
-import racing.ui.controller.RacingController
-
-fun main() {
-    RacingController().play()
-}
