+++ src/main/kotlin/racingcar/Application.kt
@@ -1,0 +1,7 @@
+package racingcar
+
+import racingcar.ui.controller.RacingController
+
+fun main() {
+    RacingController().play()
+}
