+++ src/main/kotlin/racingcar/domain/generator/anonymous/AnonymousNumberStrategy.kt
@@ -1,4 +1,4 @@
-package racing.domain.generator.anonymous
+package racingcar.domain.generator.anonymous
 
 interface AnonymousNumberStrategy {
     fun getNumber(excludeNumbers: Set<Int>): Int
