+++ src/main/kotlin/racingcar/domain/generator/trymove/TryMoveNumberGenerator.kt
@@ -1,4 +1,4 @@
-package racing.domain.generator.trymove
+package racingcar.domain.generator.trymove
 
 import kotlin.random.Random
