+++ src/main/kotlin/racingcar/domain/generator/trymove/TryMoveNumberStrategy.kt
@@ -1,4 +1,4 @@
-package racing.domain.generator.trymove
+package racingcar.domain.generator.trymove
 
 interface TryMoveNumberStrategy {
     fun getNumber(): Int
