+++ src/main/kotlin/racingcar/domain/service/CarWinnersService.kt
@@ -1,6 +1,6 @@
-package racing.domain.service
+package racingcar.domain.service
 
-import racing.domain.model.Car
+import racingcar.domain.model.Car
 
 class CarWinnersService(private val cars: List<Car>) {
     fun getWinners(): List<Car> {
