+++ src/test/kotlin/domain/CarGeneratorTest.kt
@@ -6,9 +6,9 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
-import racing.domain.generator.anonymous.AnonymousNumbersExtractor
-import racing.domain.generator.car.CarGenerator
-import racing.domain.model.Car
+import racingcar.domain.generator.anonymous.AnonymousNumbersExtractor
+import racingcar.domain.generator.car.CarGenerator
+import racingcar.domain.model.Car
 
 class CarGeneratorTest {
     private lateinit var carGenerator: CarGenerator
