+++ src/test/kotlin/domain/CarTest.kt
@@ -6,8 +6,8 @@ import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
 import org.junit.jupiter.params.provider.ValueSource
-import racing.domain.generator.trymove.TryMoveNumbersExtractor
-import racing.domain.model.Car
+import racingcar.domain.generator.trymove.TryMoveNumbersExtractor
+import racingcar.domain.model.Car
 
 class CarTest {
     @ParameterizedTest
