+++ src/test/kotlin/domain/CarWinnersServiceTest.kt
@@ -5,8 +5,8 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
-import racing.domain.model.Car
-import racing.domain.service.CarWinnersService
+import racingcar.domain.model.Car
+import racingcar.domain.service.CarWinnersService
 
 class CarWinnersServiceTest {
     private lateinit var carWinnersService: CarWinnersService
